// src/shared/api/generated/model/consentSummaryResponseDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Yoclick API
 * API multi-centro de reservas en marca blanca. Contrato canónico para la app.
 * OpenAPI spec version: 0.1.0
 */

export interface ConsentSummaryResponseDto {
  /**
     * @minimum -9007199254740991
     * @maximum 9007199254740991
     */
  clientCount: number;
  /**
     * @minimum -9007199254740991
     * @maximum 9007199254740991
     */
  privacy: number;
  /**
     * @minimum -9007199254740991
     * @maximum 9007199254740991
     */
  health: number;
  /**
     * @minimum -9007199254740991
     * @maximum 9007199254740991
     */
  marketing: number;
  /**
     * @minimum -9007199254740991
     * @maximum 9007199254740991
     */
  image: number;
  /**
     * @minimum -9007199254740991
     * @maximum 9007199254740991
     */
  parental: number;
}